+++ packages/odata-service-writer/test/index.test.ts
@@ -18,7 +18,7 @@ describe('ODataService templates', () => {
         version: OdataVersion.v2,
         metadata: '<HELLO><WORLD><METADATA></METADATA></WORLD></HELLO>',
         annotations: {
-            technicalName: 'SEPM_XYZ',
+            technicalName: '/SEPM_XYZ/SERVICE',
             xml: '<HELLO><ANNOTATION></ANNOTATION></WORLD></HELLO>'
         }
     };