+++ packages/ui5-application-writer/src/data/defaults.ts
@@ -3,6 +3,7 @@ import versionToManifestDescMapping from './version-to-descriptor-mapping.json';
 import { getUI5Libs } from './ui5Libs';
 import semVer from 'semver';
 import { t } from '../i18n';
+import { mergeObjects } from 'json-merger';
 
 /**
  * Returns a package instance with default properties.
@@ -29,6 +30,26 @@ export function packageDefaults(version?: string, description?: string): Partial
         }
     };
 }
+
+/**
+ * Merges 2 package definitions. All properties from A and from B will be present.
+ * Overlapping properties will be replaced from B. Arrays will be concatenated.
+ * `ui5.dependencies` will be de-duped.
+ *
+ * @param packageA - a partial package definition
+ * @param packageB - a partial package definition
+ * @returns - a merged package defintion
+ */
+export function mergePackages(packageA: Partial<Package>, packageB: Partial<Package>): Package {
+    const mergedPackage = mergeObjects([packageA, packageB], {
+        defaultArrayMergeOperation: 'concat'
+    });
+    // de-dup package.ui5.dependencies
+    if (mergedPackage.ui5?.dependencies) {
+        mergedPackage.ui5.dependencies = Array.from(new Set(mergedPackage.ui5.dependencies));
+    }
+    return mergedPackage;
+}
 /**
  * Returns an app instance with default properties. Every property must have a value for templating to succeed.
  *
@@ -36,13 +57,16 @@ export function packageDefaults(version?: string, description?: string): Partial
  * @returns {Partial<App>} the App instance
  */
 export function mergeApp(app: App): App {
-    // Return merged, does not update passed ref
     return Object.assign(
         {
             version: '0.0.1',
             title: t('text.defaultAppTitle', { id: app.id }),
             description: t('text.defaultAppDescription', { id: app.id }),
-            baseComponent: 'sap/ui/core/UIComponent'
+            baseComponent: 'sap/ui/core/UIComponent',
+            sourceTemplate: {
+                id: app.sourceTemplate?.id || '',
+                version: app.sourceTemplate?.version || ''
+            }
         },
         app
     ) as App;