+++ packages/ui5-config/src/middlewares.ts
@@ -75,14 +75,13 @@ export function getFioriToolsProxyMiddlewareConfig(
 }
 
 export const getMockServerMiddlewareConfig = (path?: string): CustomMiddleware<MockserverConfig> => {
-    const pathSegments = path?.split('/') || [];
+    path = path?.replace(/\/$/, ''); // Mockserver is sensitive to trailing '/'
     return {
         name: 'sap-fe-mockserver',
-        beforeMiddleware: 'fiori-tools-proxy',
+        beforeMiddleware: 'csp',
         configuration: {
             service: {
-                urlBasePath: pathSegments.slice(0, -1).join('/'),
-                name: pathSegments[pathSegments.length - 1],
+                urlPath: path || '',
                 metadataXmlPath: './webapp/localService/metadata.xml',
                 mockdataRootPath: './webapp/localService/data',
                 generateMockData: true