+++ packages/fiori-freestyle-writer/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @sap-ux/fiori-freestyle-writer
 
+## 0.11.0
+
+### Minor Changes
+
+-   574bf78: New module @sap-ux/fiori-elements-writer. Bug fix and extended api to add template info to manifest.json @sap-ux/ui5-application-writer. Add template info to manifest.json @sap-ux/fiori-freestyle-writer.
+
+### Patch Changes
+
+-   Updated dependencies [574bf78]
+    -   @sap-ux/ui5-application-writer@0.12.0
+
+## 0.10.28
+
+### Patch Changes
+
+-   Updated dependencies [893c924]
+    -   @sap-ux/ui5-application-writer@0.11.2
+
 ## 0.10.27
 
 ### Patch Changes